# Core/protocol/omni2fa.openapi.yaml
openapi: 3.1.0
info:
  title: Omni2FA API
  version: 0.6.0
  summary: Cross-stack HTTP contract for multi-method 2FA.
  description: |
    The HTTP contract that any backend implementing Omni2FA must satisfy, and
    that any frontend implementing Omni2FA expects. The same contract is the
    integration point for cross-stack combinations (React + .NET, React + Python,
    Angular + .NET, …).

    **This document is the source of truth.** Backend implementations (.NET,
    future Python) and frontend implementations (React, future Vue/Angular)
    are generated from or verified against this file.

    ## Scope per milestone

    | Milestone | Endpoints included in this spec |
    |-----------|---------------------------------|
    | v0.0 (current, pre-alpha) | Scaffolding — TOTP enrollment + login challenge, methods list/delete. |
    | v0.1 | Same as v0.0 but with implementations on .NET + React. |
    | v0.2 | + Email OTP endpoints (`/enroll/email/*`). |
    | v0.3 | + WebAuthn endpoints (`/enroll/webauthn/*`). |
    | v0.4 | + Recovery codes (`/recovery-codes/*`, returned in enrollment responses). |
    | v1.0 | Frozen — breaking changes from here on bump major. |

    See `docs/FLOWS.md` for sequence diagrams and `Core/protocol/ERROR_CODES.md`
    for the full error code catalogue.

    ## Versioning

    `info.version` matches the Omni2FA library milestone version. Packages on
    the same `MAJOR.MINOR.*` are compatible — see `docs/ROADMAP.md` "Versioning
    model".

...
servers:
...
paths:
...
  /enroll/webauthn/start:
    post:
      tags: [enroll-webauthn]
      operationId: startWebAuthnEnrollment
      summary: Begin WebAuthn enrollment — issue credential creation options.
      description: |
        Returns `PublicKeyCredentialCreationOptions` (as `optionsJson`) for the browser to pass to
        `navigator.credentials.create()`. The pending ceremony is persisted; the browser attestation
        is posted back to `/confirm`.
      security:
        - HostSession: []
      responses:
        '200':
          description: Creation options issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebAuthnEnrollStartResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: The per-user WebAuthn credential cap has been reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /enroll/webauthn/confirm:
    post:
      tags: [enroll-webauthn]
      operationId: confirmWebAuthnEnrollment
      summary: Confirm WebAuthn enrollment by verifying the browser attestation.
      security:
        - HostSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebAuthnEnrollConfirmRequest'
      responses:
        '200':
          description: WebAuthn credential enrolled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MethodCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Attestation failed verification, or no pending enrollment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No pending WebAuthn enrollment for this user, or it has expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ------------------------------------------------------------------
  #  enrollment — TOTP
  # ------------------------------------------------------------------
...
components:
...
  securitySchemes:
    HostSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The host application's full session token. Omni2FA does not issue
        this — your existing auth does. Endpoints under `methods` and
        `enroll-*` require this. The user has fully logged in.
    PreAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived (default 3-5 min) pre-auth token issued by Omni2FA after
        the host has verified the user's password but before 2FA verification
        completes. Used exclusively by `/challenge/*` endpoints.
        Token is invalidated immediately after a successful `/challenge/verify`.

  # ------------------------------------------------------------------
  #  Parameters
  # ------------------------------------------------------------------
...
  schemas:
...
    WebAuthnEnrollStartResponse:
      type: object
      required: [enrollmentId, optionsJson]
      properties:
        enrollmentId:
          type: string
          format: uuid
          description: Identifier for the pending ceremony. Pass back in `confirm`.
        optionsJson:
          type: string
          description: |
            `PublicKeyCredentialCreationOptions` JSON for `navigator.credentials.create()`.
            Challenge, user id, and excluded credential ids are base64url-encoded.
    WebAuthnEnrollConfirmRequest:
      type: object
      required: [enrollmentId, attestationResponseJson]
      properties:
        enrollmentId:
          type: string
          format: uuid
          description: Value returned from `/enroll/webauthn/start`.
        attestationResponseJson:
          type: string
          description: The JSON produced by `navigator.credentials.create()`, base64url-encoded fields.
        name:
          type: [string, "null"]
          description: Optional human-readable label for this credential (e.g. "YubiKey 5").

    # --- TOTP enrollment ---
...
    MethodCreatedResponse:
      type: object
      required: [methodId]
      properties:
        methodId:
          type: string
          format: uuid
          description: Identifier of the newly created method.
        recoveryCodes:
          type: [array, "null"]
          items:
            type: string
          description: |
            Present only when this enrollment generated recovery codes (i.e. it was the user's
            first method). Plaintext, shown once — the frontend must display and have the user
            save them. Null on every subsequent enrollment.
...
    ErrorResponse:
      type: object
      required: [code, message]
      description: |
        Uniform error envelope returned by every non-2xx response. Frontends
        switch on `code` (machine-readable, stable). `message` is informational
        and may be localized.
        See `Core/protocol/ERROR_CODES.md` for the full catalogue.
      properties:
        code:
          type: string
          description: Machine-readable error code, stable across versions.
          example: 'INVALID_CODE'
        message:
          type: string
          description: Human-readable message. May be localized by host.
          example: 'The code you entered is invalid.'
        details:
          type: [object, "null"]
          additionalProperties: true
          description: Optional structured context. Shape depends on `code`.

  # ------------------------------------------------------------------
  #  Reusable responses
  # ------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid session token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...